Refactor react-redux provider
